Refactor dashboard components: remove unused ProfileDropdown, ProgressCard, RecentActivity, and UpcomingDeadlines; introduce new DashboardHeader, DashboardTabs, DashboardCards, DashboardCalendar, DashboardNotifications, and ProfileSidebar components for improved structure and functionality.
