+++ app/(protected)/dashboard/components/DashboardHeader.tsx
@@ -1,0 +1,10 @@
+export default function DashboardHeader() {
+  return (
+    <div className="mb-8">
+      <h1 className="text-4xl font-bold text-gray-600 mb-12 mt-15 leading-[3rem]">
+        Everything You Need for Career, Civic,  <br /> and Life Readiness is{" "}
+        <span className="italic text-adult-green">Here.</span>
+      </h1>
+    </div>
+  )
+}
